requirements.txt
pymysql~=1.1.1
paramiko~=3.5.0
sshtunnel~=0.4.0
pytest~=8.3.3
pytest-html
pytest-xdist
pytest-rerunfailures
pytest-sugar
pytest-ordering
allure-pytest
pysftp==0.2.8
requests~=2.32.3
mock
docopt
selenium~=4.25.0
docker
docker-compose
xlrd
xlutils
xlwt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
regex
bs4
testrail~=0.3.15
pytest-testrail
opencv-python
scikit-image
flaky
mysql-connector-python~=9.0.0
jumpssh~=1.6.5
pymongo~=4.10.1
pyautogui~=0.9.54
py~=1.11.0
pytz~=2024.2
jsonschema~=4.23.0